fix for issue 41125
